[FE] home의 현재 클릭한 사물함 recoil 초기화 기능 main으로 이동 #849
