+++ frontend/src/pages/MainPage.tsx
@@ -1,12 +1,31 @@
-import { useRef, useState } from "react";
+import { useEffect } from "react";
 import styled from "styled-components";
 import { currentFloorSectionState } from "@/recoil/selectors";
-import { useRecoilState, useRecoilValue } from "recoil";
+import { useRecoilState, useRecoilValue, useResetRecoilState } from "recoil";
 import { currentSectionNameState } from "@/recoil/atoms";
+import { currentCabinetIdState, targetCabinetInfoState } from "@/recoil/atoms";
+import useMenu from "@/hooks/useMenu";
 import SectionPaginationContainer from "@/components/SectionPagination/SectionPagination.container";
 import CabinetListContainer from "@/components/CabinetList/CabinetList.container";
 
 const MainPage = () => {
+  const { closeAll } = useMenu();
+
+  const resetTargetCabinetInfo = useResetRecoilState(targetCabinetInfoState);
+  const resetCurrentCabinetId = useResetRecoilState(currentCabinetIdState);
+
+  useEffect(() => {
+    closeAll();
+    resetTargetCabinetInfo();
+    resetCurrentCabinetId();
+
+    return () => {
+      closeAll();
+      resetTargetCabinetInfo();
+      resetCurrentCabinetId();
+    };
+  }, []);
+
   let touchStartPosX: number;
 
   const sectionList = useRecoilValue<Array<string>>(currentFloorSectionState);
